requirements.txt
requests
pandas
matplotlib
kivy[full]
kivy_examples
webview-android
kivy_garden
pywebview
python-telegram-bot
#config.yml (remove later)
opencv-python
pytesseract
mysql
plotly
mysql-connector-python
sqlalchemy
pymysql
pyyaml
yahoo_fin
yfinance
ipywidgets
django
djangorestframework






